test: 테스트 코드 위치 리팩터링
